refactor: move delete db form component in dir structure
