+++ app/components/App/App.js
@@ -5,7 +5,7 @@ import ExplorerContainer from '../Explorer/Explorer';
 import ConnectionFormContainer from '../modals/ConnectionForm';
 import ConfirmRowDelete from '../modals/ConfirmRowDelete';
 import DatabaseForm from '../modals/DatabaseForm';
-import DeleteDatabaseForm from '../modals/DeleteDatabaseForm';
+import DeleteDatabaseForm from '../DeleteDatabaseForm/DeleteDatabaseForm';
 import DeleteTableForm from '../modals/DeleteTableForm';
 import TableForm from '../modals/TableForm';
 import DevTools from '../DevTools';
